[feat] Improve RequestHandler and EndpointOutput type declarations (#1778)
Created DefaultBody type, which can be used as the default for Request or Response bodies. Added a generic parameter to `RequestHandler` and an generic parameter to `EndpointOutput`. Together, these let you specify the type of the body of the Response you want to return from a RequestHandler function. The default of both of these parameters is `DefaultBody`.