[docs] note about host specific SPA redirect config (#6745)
An often recurring question using the static adapter in SPA mode is why it keeps seeing "file not found" after being deployed.  This confusion stems from the idea that SvelteKit can somehow handle the redirects necessary.  This change adds a small explainer text to remind developers that this is something they have to configure themselves as it fully depends on what kind of server they use.