docs: update SW docs to use paths as cache-keys (#9718)
For some Reason Chrome and other Chromium based browsers do not match a cached response when using a request as the cache-key if the cache was initialized from a path and not a request object. The example uses an array of pathnames to populate the cache, so you also need to use paths to retrieve responses from the cache. This weird behaviour causes cache-misses on resources that are very clearly in the cache.

Using the pathname itself as the cache key fixes the issue.

(this was a pain to figure out)