feat(node): add `PORT_HEADER` env var for reverse proxies with non-standard ports (#11249)
* feat: add `PORT_HEADER` env var for reverse proxies with non-standard ports

Add support for the `PORT_HEADER` environment variable, that can be used
to read the original port of a request from a reverse proxy, which is
needed if the reverse proxy is hosted on a non-standard port
(e.g. not port 80 for http nor port 443 for https).

This port is normally added as part of a [`x-forwarded-port`][1], which
is used by many reverse proxies, although it's not as popular as the
more common [`x-forwarded-proto`][2] and [`x-forwarded-host`][3]
headers.

[1]: https://docs.aws.amazon.com/elasticloadbalancing/latest/classic/x-forwarded-headers.html#x-forwarded-port
[2]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-Proto
[3]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-Host

* fixup! feat: add `PORT_HEADER` env var for reverse proxies with non-standard ports

Remove link to AWS docs.